project user error render

+++ database/migrations/2025_04_29_152035_create_project_user_table.php
@@ -11,13 +11,15 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('project_user', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('project_id')->constrained()->onDelete('cascade');
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->unique(['project_id', 'user_id']);
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('project_user')) {
+            Schema::create('project_user', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('project_id')->constrained()->onDelete('cascade');
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->unique(['project_id', 'user_id']);
+                $table->timestamps();
+            });
+        }
     }
 
     /**
